Implement same behaviour as before but using a modified shadcn Pagination.
Why modified? Because the shadcn version uses `a` tag link to navigate from page to page. Not sure if we will need to naviguate pages in the future but the actual implementation only use `useState` to naviguate.

This is a first step in order to implement a more advanced pagination with the ability to choose the page more easily and also choose the page size. Ideally the page size will be stored in localStorage by context.